requirements.txt
Django==1.3.1
PIL==1.1.7
argparse==1.2.1
django-piston==0.2.3
mongoengine==0.5.2
pymongo==2.1.1
wsgiref==0.1.2
